message

+++ project_wings/src/main/java/com/wings/mywiki/controller/AdminController.java
@@ -38,22 +38,24 @@ public class AdminController {
 		return reportMap;
 	}
 	
-	//신고 접수
+	//신고 승인 & 거부
 	@PostMapping("approveReport")
 	@ResponseStatus(HttpStatus.OK)
 	public HashMap<String, Object> approveReport(@RequestBody HashMap<String, Object> map) {
-		//int reportUserId = adminService.getReportUserId(map.get("reportId"));
-		int reportedUserId = adminService.getReportedUserId((int)map.get("reportId"));
+		int approve=(int) map.get("approve");	//1:승인, 0:거절
 		
-		// 신고 된 사람의 신고 된 횟수 +1
-		adminService.updateReportedNum(reportedUserId);
-		
-		//신고 된 횟수 가져 옴
-		int reportedNum = adminService.getReportedNum(reportedUserId);
-		if (reportedNum>=5) {	//신고 된 횟수가 5회 이상이면 탈퇴 조취
-			adminService.deleteUser(reportedUserId);
+		if (approve==1) {
+			int reportedUserId = adminService.getReportedUserId((int)map.get("reportId"));
+			
+			// 신고 된 사람의 신고 된 횟수 +1
+			adminService.updateReportedNum(reportedUserId);
+			
+			//신고 된 횟수 가져 옴
+			int reportedNum = adminService.getReportedNum(reportedUserId);
+			if (reportedNum>=5) {	//신고 된 횟수가 5회 이상이면 탈퇴 조취
+				adminService.deleteUser(reportedUserId);
+			}
 		}
-		
 		//신고 리스트에서 삭제
 		adminService.deleteReport((int) map.get("reportId"));
 		
@@ -64,20 +66,6 @@ public class AdminController {
 		return reportMap;
 	}
 	
-	//신고 접수 거부
-	@PostMapping("rejectReport")
-	@ResponseStatus(HttpStatus.OK)
-	public HashMap<String, Object> rejectReport(@RequestBody HashMap<String, Object> map) {
-		//신고 리스트에서 삭제
-		adminService.deleteReport((int) map.get("reportId"));
-		
-		HashMap<String, Object> reportMap = new HashMap<String, Object>();
-		List<ReportVO> reportList = adminService.getAllReports();
-		reportMap.put("reportList", reportList);
-
-		return reportMap;
-	}
-
 	// 공지사항 등록
 	@PostMapping(value = "insert")
 	@ResponseStatus(HttpStatus.CREATED)
